Add g.insertConstant(val)
Replaces insertConstant(g, val), and makes it more consistent with other
node construction function.

Moves SourceRange to its own header so that we don't have to include
the entire lexer in ir.h
