+++ torch/csrc/jit/constants.h
@@ -1,8 +1,6 @@
 #pragma once
-#include "ATen/ATen.h"
 #include "torch/csrc/jit/ivalue.h"
-#include "torch/csrc/jit/ir.h"
-#include "torch/csrc/jit/script/lexer.h"
+#include "torch/csrc/jit/source_range.h"
 #include "torch/csrc/WindowsTorchApiMacro.h"
 
 // helpers for handling constants in the IR
@@ -10,10 +8,16 @@
 // - implement primitive constant ops.
 namespace torch { namespace jit {
 
+struct Graph;
+struct Value;
+
+// note: prefer g.insertConsant(val, loc) which does exactly the same thing
+// this function is only declared/defined here because its implementation is
+// closely related to the implementation of prim::Constant that is also in constants.cpp
 TORCH_API Value* insertConstant(
     Graph& g,
     IValue val,
-    at::optional<script::SourceRange> loc = at::nullopt);
+    at::optional<SourceRange> loc = at::nullopt);
 
 
 //////////////////////////////////////////////////////////////////////////////////
