+++ torch/csrc/jit/named_value.h
@@ -1,17 +1,17 @@
 #pragma once
 #include "ATen/ATen.h"
 #include "torch/csrc/jit/ir.h"
-#include "torch/csrc/jit/script/tree.h"
+#include "torch/csrc/jit/source_range.h"
 
 namespace torch { namespace jit {
 
 struct NamedValue {
-  NamedValue(const script::SourceRange& loc, const std::string& name, Value* value)
+  NamedValue(const SourceRange& loc, const std::string& name, Value* value)
   : loc(loc), name(name), value(value) {}
-  NamedValue(const script::SourceRange& loc, int i, Value* value)
+  NamedValue(const SourceRange& loc, int i, Value* value)
   : loc(loc), name("argument " + std::to_string(i)), value(value) {}
 
-  script::SourceRange loc;
+  SourceRange loc;
   std::string name;
   Value* value;
 };
