+++ torch/csrc/jit/passes/erase_number_types.cpp
@@ -16,7 +16,7 @@ static void EraseNumberTypesOnBlock(Block* block) {
         if(it->output()->type()->isSubtypeOf(NumberType::get())) {
           auto s = *constant_as<at::Scalar>(it->output());
           WithInsertPoint guard(*it);
-          Value* r = insertConstant(*block->owningGraph(), s.toTensor());
+          Value* r = block->owningGraph()->insertConstant(s.toTensor());
           it->output()->replaceAllUsesWith(r);
         }
       } break;
