+++ torch/csrc/jit/passes/loop_unrolling.cpp
@@ -145,7 +145,7 @@ Value* intMath(Symbol sym, Value* a, Value* b) {
       ->setType(IntType::get());
 }
 Value* intMath(Symbol sym, Value* a, int64_t b) {
-  return intMath(sym, a, insertConstant(*a->owningGraph(), b));
+  return intMath(sym, a, a->owningGraph()->insertConstant(b));
 }
 
 // Replaces the builtin loop counter with a "mutable" variable outside of the loop.
@@ -153,7 +153,7 @@ void replaceLoopCounter(Node *loop) {
   Graph *graph = loop->owningGraph();
   Block *body = loop->blocks().at(0);
   WithInsertPoint guard(loop);
-  Value* init_counter = insertConstant(*graph, 0);
+  Value* init_counter = graph->insertConstant(0);
 
   loop->insertInput(2, init_counter);
   loop->insertOutput(0);
