+++ torch/csrc/jit/tracer.h
@@ -121,7 +121,7 @@ inline Value* getValueTrace(const Variable& var) {
   auto & value_map = getTracingState()->value_map;
   auto it = value_map.find(var);
   if (it == value_map.end()) {
-    Value *constant = insertConstant(*state->graph, var.data());
+    Value *constant = state->graph->insertConstant(var.data());
     constant->inferTypeFrom(var.data());
     it = value_map.emplace_hint(it, var, constant);
   }
